+++ docs/examples/hello-world/src/Application.ts
@@ -1,16 +1,18 @@
+import { Application as DillPixelApplication, TransitionType } from 'dill-pixel';
 import { MyScreen } from './state/MyScreen';
 import { LoadScreen } from './state/LoadScreen';
-import { Application as DillPixelApplication } from '@relish-studios/dill-pixel';
 
 export class Application extends DillPixelApplication {
-  public get defaultState() {
-    return MyScreen.NAME;
-  }
+
+  // Register the load screen and define the transition type between states
 
   protected setup() {
     (globalThis as any).__PIXI_APP__ = this;
     this.registerDefaultLoadScreen(LoadScreen);
-  }  
+    this.state.defaultTransitionType = TransitionType.TRANSITION_SIMPLE_INTERSTITIAL;
+  }
+
+  // Register the states
 
   protected registerStates(): void {
     this.state.register(MyScreen);