+++ docs/examples/hello-world/src/state/LoadScreen.ts
@@ -1,15 +1,20 @@
-import { LoadScreen as DillPixelLoadScreen } from '@relish-studios/dill-pixel';
-import { Sprite, Point } from 'pixi.js';
+import { LoadScreen as DillPixelLoadScreen } from 'dill-pixel';
+import { Point, Sprite } from 'pixi.js';
+
+// The load screen is simply a black background
 
 export class LoadScreen extends DillPixelLoadScreen {
-  public static NAME: string = 'LoadScreen';
-
+  static NAME: string = 'LoadScreen';
   private _bg!: Sprite;
+
+  // Create the black background
 
   public init(pSize: Point) {
     super.init(pSize);
     this._bg = this.add.coloredSprite(0x000000);
   }
+
+  // When the screen resizes, resize the background
 
   public onResize(pSize: Point) {
     super.onResize(pSize);